pyproject.toml
[project]
name = "spite"
version = "0.1.0"
description = "Automated job hunting for people who know the market is broken."
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "typer>=0.13.0",
    "rich>=13.9.0",
    "sqlalchemy>=2.0.0",
    "playwright>=1.48.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "google-genai>=1.72.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
]

[project.scripts]
spite = "spite.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["spite"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]
